requirements.txt
tkinter
tkcalendar
pandas
sqlalchemy
openpyxl
pillow
pyodbc